+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.7.0] - 2023-05-04
+
+### New
+
+- Added new type of giver - MsigV2
+ 
 ## [1.6.1] - 2023-03-28
 
 ### New